+++ backend/garpix_notify/utils/__init__.py
@@ -2,3 +2,4 @@ from .file import get_file_path
 from .sms_checker import SMSClient
 from .receiving import receiving_users
 from .call_code_cheker import CallClient
+from .send_data import notify_create